Refactor DNS service
